Исправлен баг с переполнением счетчика количества объектов узла. Исправлен баг с некорректным добавлением объекта обратно в дерево при переходе в новый узел для сжатого квадродерева. Исправлен баг с некорректным обновлением индекса добавленного объекта в дерево для квадродерева с пропусками (ранее не учитывалось, что для временно пропавшего объекта возможна потеря экземпляра при добавлении другого объекта в тот же узел до того, как пропавший объект будет возвращен в дерево); аналогичный подход применен и к логике удаления объекта, если он помечен как пропавший (что, правда, маловероятно). Рефакторинг.

+++ Assets/SpatialPartitionSystem/Core/Scripts/Series/Common/Node.cs
@@ -6,7 +6,7 @@
     {
         public int parentIndex;
         public int firstChildIndex;
-        public byte objectsCount;
+        public int objectsCount;
         public bool isLeaf;
         
         public byte depth;
